requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
httpx==0.27.0
# prisma==0.11.0             # если буду использовать prisma-client-py
# openai>=1.30.1
httpx==0.27.0